Squashed commit of the following:
