+++ src/main/java/org/braekpo1nt/mctmanager/games/game/parkourpathway/states/InitialState.java
@@ -1,0 +1,8 @@
+package org.braekpo1nt.mctmanager.games.game.parkourpathway.states;
+
+import org.braekpo1nt.mctmanager.games.experimental.DoNothingState;
+import org.braekpo1nt.mctmanager.games.game.parkourpathway.ParkourParticipant;
+import org.braekpo1nt.mctmanager.games.game.parkourpathway.ParkourTeam;
+
+public class InitialState implements ParkourPathwayState, DoNothingState<ParkourParticipant, ParkourTeam> {
+}
