+++ src/main/java/org/braekpo1nt/mctmanager/games/game/parkourpathway/states/ParkourPathwayState.java
@@ -1,0 +1,8 @@
+package org.braekpo1nt.mctmanager.games.game.parkourpathway.states;
+
+import org.braekpo1nt.mctmanager.games.experimental.GameStateBase;
+import org.braekpo1nt.mctmanager.games.game.parkourpathway.ParkourParticipant;
+import org.braekpo1nt.mctmanager.games.game.parkourpathway.ParkourTeam;
+
+public interface ParkourPathwayState extends GameStateBase<ParkourParticipant, ParkourTeam> {
+}
